% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_states.R
\name{assign_states}
\alias{assign_states}
\title{Assign Initial States}
\usage{
assign_states(x, mix.mat = NULL, thresholds = NULL)
}
\arguments{
\item{x}{The observed count vector.}

\item{mix.mat}{The parameter matrix output by fit_nb_mix() or fit_pois_mix().}

\item{thresholds}{A vector of thresholds that define the different states. An alternative to supplying mix.mat.}
}
\description{
Initial HMM state assignments for a vector of counts based on defined thresholds. States are assigned such that the state above the highest threshold is State 1, the state above the next highest threshold is State 2, etc.
}
