% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bound_check.R
\name{bound_check}
\alias{bound_check}
\title{Genome Boundary Check}
\usage{
bound_check(fragments, genome.obj)
}
\arguments{
\item{fragments}{The random fragments being tested.}

\item{genome.obj}{The name of the genome object of interest.}
}
\value{
A vector of index values for terms in violation of genomic boundaries.
}
\description{
Check that simulated random fragment ranges do not violate genome/chromosome boundaries. Not intended to be run alone. Called within make_fragments().
}
\examples{
\dontrun{
bound_check(fragments = frag.object, genome.obj = genome.object)
}

}
