% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_reference_model.R
\name{build_reference_model}
\alias{build_reference_model}
\title{Build Reference Model}
\usage{
build_reference_model(xint.obj)
}
\arguments{
\item{xint.obj}{An xInt object. The output of make_xint_dataset().}
}
\description{
Build feature-wise negative binomial models for the reference or control samples.
}
