% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_objects.R
\name{bulk_objects}
\alias{bulk_objects}
\title{Make xInt Objects in Batch}
\usage{
bulk_objects(
  sites,
  conditions,
  condition.levels,
  features.list,
  type = c("overlap", "nearest"),
  feature.names = NULL,
  min.overlap = 1,
  id.col = "name",
  ric.dat = NULL
)
}
\arguments{
\item{conditions}{The condition for each dataset. Must be same length as sites.}

\item{condition.levels}{The factor levels of conditions.
All entries should be present in conditions.
Length should be the number of unique conditions.}

\item{features.list}{The list containing to different feature-sets.}

\item{type}{The type of xInt object to construct.
Can be "overlap" for xIntOverlap (default) or "nearest" for xIntNearest.}

\item{feature.names}{A character vector of names for each feature in features.list.
When NULL (default), features.list must be named.}

\item{min.overlap}{The minimum amount of overlap to count two regions as overlapping.
Defaults to 1.}

\item{id.col}{The ID column in the features GRanges object. Default "name".}

\item{ric.dat}{A separate SiteList object for random integration control sites.
Can be NULL (default).}

\item{site}{A SiteList object.}
}
\value{
A list of xInt objects.
}
\description{
Make xIntOverlap or xIntNearest objects in bulk from a list of feature-sets.
}
