% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw.R
\name{bw}
\alias{bw}
\title{Baum-Welch algorithm}
\usage{
bw(object, control = bwcontrol(), ...)
}
\arguments{
\item{object}{An object of class "dthmm", "mmglm0", "mmglm1", "mmglmlong1", or "mmpp".}

\item{control}{A list of control settings for the iterative process. These can be changed by using the function bwcontrol.}
}
\description{
A modified version of the BaumWelch.dthmm() function in the HiddenMarkov package. This function is modified to use the absolute relative difference as the convergence criteria instead of absolute difference. This is intended to make the function more capable of dealing with relatively flat likelihood functions. See the original package manual for more information on its inner-workings (https://cran.r-project.org/web/packages/HiddenMarkov/HiddenMarkov.pdf).
}
