% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dG_internal.R
\name{calc_dG_internal}
\alias{calc_dG_internal}
\title{calculate_dG Internal Function}
\usage{
calc_dG_internal(seqs, vals, len)
}
\arguments{
\item{seqs}{The sequences of interest.}

\item{vals}{The deltaG values for each trimeric unit.}

\item{len}{The length of each sequence.}
}
\value{
A numeric vector of row means.
}
\description{
Called within calculate_dG().
Performs the dG calculations across the provided sequences.
}
