% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cpm.R
\name{calculate_cpm}
\alias{calculate_cpm}
\title{Calculate CPM Values}
\usage{
calculate_cpm(y, lib.size, log = TRUE, prior.count = 2, lib.prior = NULL)
}
\arguments{
\item{y}{Input count values}

\item{lib.size}{A numeric vector of library sizes of each sample.}

\item{log}{Boolean. Whether or not to return log-CPM. Default TRUE.}

\item{prior.count}{The constant added to count values to avoid taking log2(0).}

\item{lib.prior}{The constant added to library size when calculating log-CPM.
When NULL, lib.prior = prior.count.}
}
\value{
A matrix or vector of log-CPM values
}
\description{
Perform CPM normalization on count values.
}
