% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_dG.R
\name{calculate_dG}
\alias{calculate_dG}
\title{Calculate B-to-A deltaG}
\usage{
calculate_dG(
  site.list,
  step.len = 50,
  genome.obj,
  current.start = 1,
  tsd = 5,
  collapse = TRUE
)
}
\arguments{
\item{site.list}{The list or GRangesList containing the mapped site coordinates.}

\item{step.len}{The target length of the returned vector of dG values.}

\item{genome.obj}{The genome object of interest.}

\item{current.start}{The position in the target site duplication currently described by start.
This is used for centering the site coordinates.}

\item{tsd}{The total length of the target site duplication.
This is used for centering the site coordinates.}

\item{collapse}{Boolean. Whether or not to collapse the output into a single data.frame. Defaults to TRUE.}
}
\value{
A list of data frames containing the sequence position and the corresponding dG value for each element in site.list.
}
\description{
Calculates the free energy change upon B to A transition of DNA.
Relies on trimeric dG values calculated by Tolstorukov, et al. (DOI: 10.1016/S0006-3495(01)75973-5).
Utilizes a 3 bp sliding window with a 1 bp step across the sequence of interest.
}
\examples{
library(BSgenome.Hsapiens.UCSC.hs1)
data(sites2)
calculate_dG( site.list = sites2, genome.obj = Hsapiens )

}
