% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sites.R
\name{check_sites}
\alias{check_sites}
\title{Check Sites for Compatibility}
\usage{
check_sites(site.list)
}
\arguments{
\item{site.list}{The list of GRanges objects or GRangesList containing the mapped site coordinates.}
}
\value{
None
}
\description{
Verify that imported IS datasets meet the requirements for downstream xInt analyses.
}
