% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_estimates.R
\name{clonal_estimates}
\alias{clonal_estimates}
\title{Clonal Statistics of Integration Site Datasets}
\usage{
clonal_estimates(sites, conditions = NULL, score.col = NULL)
}
\arguments{
\item{sites}{A SiteList object}

\item{conditions}{A character vector of conditions corresponding to the individual datasets.
Must be equal to length(sites) and must be in the appropriate order.
When not NULL (default), replicates from the same condition are pooled.}

\item{score.col}{The column in the SiteList object holding site count values.
This is included for instances when e.g., site abundance is estimated by MLE methods.
When not NULL, this column should be present in the SiteList object.}
}
\value{
A data frame containing the Gini coefficient, Shannon index, Pielou index, and Simpson index for each sample/condition.
}
\description{
Clonal Statistics of Integration Site Datasets
}
