% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonal_sites.R
\name{clonal_sites}
\alias{clonal_sites}
\title{Define Clonal Sites}
\usage{
clonal_sites(sites, conditions, min.threshold = 0, score.col = NULL)
}
\arguments{
\item{sites}{A SiteList object}

\item{conditions}{A character vector of conditions corresponding to the individual datasets.
Must be equal to length(sites) and must be in the appropriate order.
When not NULL (default), replicates from the same condition are pooled.}

\item{min.threshold}{The minimum threshold for a site to be considered clonal, expressed as a fraction of all sites
or as a fraction of the total estimated abundance. Defaults to 0 (all sites are returned).}

\item{score.col}{The column in the SiteList object holding site count values.
This is included for instances when e.g., site abundance is estimated by MLE methods.
When not NULL, this column should be present in the SiteList object.}
}
\value{
A GRanges object
}
\description{
Define Clonal Sites
}
