% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_features.R
\name{collapse_features}
\alias{collapse_features}
\title{Collapse Overlapping Features}
\usage{
collapse_features(
  features,
  ignore.strand = TRUE,
  original.names = FALSE,
  id.col = NULL
)
}
\arguments{
\item{features}{A GRanges object holding the features-of-interest.}

\item{ignore.strand}{Logical. Whether or not to ignore the strandedness of the features. Defaults to TRUE.}

\item{original.names}{Logical. Whether or not keep the original feature names. In the case of overlapping features, the feature names are changed to include them all, separated by hyphens. This can be slow. If FALSE, new features are named Feature 1 to Feature N.}

\item{id.col}{The metadata column containing feature names. Must be provided if original.names = TRUE. Can be NULL.}
}
\value{
A GRanges object of collapsed features.
}
\description{
Collapse features-of-interest to eliminate overlapping features that might result in multiple counting of individual integration sites.
}
