% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_density.R
\name{compare_density}
\alias{compare_density}
\title{Compare Feature Density Between Conditions}
\usage{
compare_density(x, p.adjust.method = "BH", method = c("t.test", "anova"))
}
\arguments{
\item{x}{The output of feature_density() with options average = TRUE and conditions are given.}

\item{p.adjust.method}{Method for p-value adjustment. Default is "BH" (Benjamini-Hochberg).}

\item{method}{Statistical test to use. Either 't.test' (default) or 'anova'.}
}
\value{
A data frame containing the results of statistical comparisons between conditions.
}
\description{
Performs statistical tests to compare feature density between different conditions.
}
