% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_stats.R
\name{contrast_stats}
\alias{contrast_stats}
\title{Extract Comparison Statistics}
\usage{
contrast_stats(
  x,
  contrast = Inf,
  n.hits = Inf,
  p.cutoff = 0.05,
  fc.cutoff = 1,
  pc.cutoff = 0,
  p.adjust.method = "BH",
  sort.by = "p.adj",
  decreasing = FALSE,
  significant.only = FALSE
)
}
\arguments{
\item{x}{The output from overlap_comparisons() or single_overlap_comparisons()}

\item{contrast}{The name or numeric index of the contrast to extract.
If Inf, results for all contrasts will be returned in a single dataframe.}

\item{n.hits}{Number of top hits to extract. Default is Inf (all hits).}

\item{p.cutoff}{P-value cutoff for filtering results. Default is 0.05.}

\item{fc.cutoff}{Log2 fold change cutoff for filtering results. Default is 1.
Only applied when contrast is specified.}

\item{pc.cutoff}{Percent change cutoff for filtering results. Default is 0.
Only applied when contrast is specified.}

\item{p.adjust.method}{Method for p-value adjustment. Default is "BH" (Benjamini-Hochberg).}

\item{sort.by}{Column to sort results by. Default is "p.adj".}

\item{decreasing}{Boolean indicating if sorting should be in decreasing order. Default is FALSE.}

\item{significant.only}{Boolean indicating whether or not to only return significant features. Defaults to FALSE.}
}
\value{
A data frame with rows for each feature and columns for various statistics
}
\description{
Neatly extract p-value and effect size information for specified contrasts from the
outputs of overlap_comparisons(), single_overlap_comparisons(), or wilcox_overlaps().
}
\examples{
data(xobj)
glm.results <- overlap_comparisons(xint.obj = xobj)

contrast_stats(
  glm.results,
  contrast = 1,
  n.hits = 2
  p.cutoff = 1
 )

}
