% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_comparison.R
\name{cross_comparison}
\alias{cross_comparison}
\title{Calculate Cross-Sample Overlap}
\usage{
cross_comparison(
  sites,
  overlap.coef = c("overlap", "jaccard", "dice", "none"),
  return.sites = FALSE
)
}
\arguments{
\item{sites}{A SiteListObject object.}

\item{overlap.coef}{One of 'overlap' (Szymkiewicz–Simpson overlap coefficient),
'jaccard' (Jaccard index), 'dice', (Dice-Sørensen coefficient), or
'none' (the number of overlapping sites is returned).}

\item{return.sites}{Returns a GRanges object holding the sites present in 2 or more datasets.}
}
\value{
A matrix of counts denoting the number of overlapping sites between each sample.
}
\description{
Calculates the number of overlapping sites between each sample in a SiteListObject.
}
\examples{
cross_comparison(sites = sites)

}
