% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estep.R
\name{estep}
\alias{estep}
\title{Expectation Algorithm}
\usage{
estep(x, Pi, delta, distn, pm, pn = NULL)
}
\arguments{
\item{x}{is a vector of length n containing the univariate observed process. Alternatively, x could be specified as NULL, meaning that the data will be added later (e.g. simulated).}

\item{Pi}{is the m × m transition probability matrix of the homogeneous hidden Markov chain}

\item{delta}{is the marginal probability distribution of the m hidden states at the first time point}

\item{distn}{is a character string with the abbreviated distribution name. Distributions provided by the package are Beta ("beta"), Binomial ("binom"), Exponential ("exp"), GammaDist ("gamma"), Lognormal ("lnorm"), Logistic ("logis"), Normal ("norm"), and Poisson ("pois"). See topic Mstep, Section “Modifications and Extensions”, to extend to other distributions.}

\item{pm}{is a list object containing the (Markov dependent) parameter values associated with the distribution of the observed process.}
}
\description{
A modified version of the Estep() function in the HiddenMarkov package. The function uses a modified forward-backward algorithm (fb(), instead of the HiddenMarkov forwardback()). Parameter definitions are taken directly from the HiddenMarkov manual (https://cran.r-project.org/web/packages/HiddenMarkov/HiddenMarkov.pdf).
}
