% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_features.R
\name{expand_features}
\alias{expand_features}
\title{Expand Feature Coordinates}
\usage{
expand_features(
  features,
  upstream = 0,
  downstream = 0,
  genome.obj = NULL,
  ignore.strand = FALSE
)
}
\arguments{
\item{features}{The GRanges object holding the features-of-interest}

\item{upstream}{The amount to expand (in bp) upstream of the FOI.
Features with strand = '*' are treated like strand = '+'.}

\item{downstream}{The amount to expand (in bp) downstream of the FOI.}

\item{genome.obj}{The genome object of interest. This is used for checking the validity of
the expanded coordinates. Can be NULL.}

\item{ignore.strand}{Boolean. Whether or not to ignore strandedness. In this case, all features are treated like strand = '+'.
Defaults to FALSE}
}
\value{
A GRanges object containing the expanded coordinates.
}
\description{
Expands feature coordinates by a defined amount.
}
