% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb.R
\name{fb}
\alias{fb}
\title{Forward-Backward Algorithm}
\usage{
fb(x, Pi, delta, distn, pm, pn = NULL, fortran = TRUE)
}
\arguments{
\item{x}{is a vector of length n containing the univariate observed process. Alternatively, x could be specified as NULL, meaning that the data will be added later (e.g. simulated).}

\item{Pi}{is the m × m transition probability matrix of the homogeneous hidden Markov chain}

\item{delta}{is the marginal probability distribution of the m hidden states at the first time point}

\item{distn}{is a character string with the abbreviated distribution name. Distributions provided by the package are Beta ("beta"), Binomial ("binom"), Exponential ("exp"), GammaDist ("gamma"), Lognormal ("lnorm"), Logistic ("logis"), Normal ("norm"), and Poisson ("pois"). See topic Mstep, Section “Modifications and Extensions”, to extend to other distributions.}

\item{pm}{is a list object containing the (Markov dependent) parameter values associated with the distribution of the observed process.}

\item{pn}{is a list object containing the observation dependent parameter values associated with the distribution of the observed process.}

\item{fortran}{logical, if TRUE (default) use the Fortran code, else use the R code.}
}
\description{
A modified version of the forwardback() function in the HiddenMarkov package. This modified function was first written by Reed A. Cartwright (https://gist.github.com/reedacartwright/7b5339862c9abd388e10b101bfed6f07). The modifications help prevent numeric underflow by centering probability calculations around probmax. See the HiddenMarkov manual for more information on the stock function (https://cran.r-project.org/web/packages/HiddenMarkov/HiddenMarkov.pdf).
}
