% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_density.R
\name{feature_density}
\alias{feature_density}
\title{Feature Density}
\usage{
feature_density(
  site.list,
  features,
  genome.obj,
  win.size = 1e+06,
  min.overlap = 1L,
  average = TRUE,
  remove.outliers = FALSE
)
}
\arguments{
\item{site.list}{The list or GRangesList containing the mapped site coordinates.}

\item{features}{The feature-set of interest of interest. Should be a GRanges object.}

\item{genome.obj}{The genome object of interest.}

\item{win.size}{The total distance around each site.}

\item{min.overlap}{The minimum amount of overlap required to call two regions overlapping. Defaults to 1L.}

\item{average}{Boolean. Whether or not to average the number of features within win.size for each dataset. Defaults to TRUE.}

\item{remove.outliers}{Whether or not to remove regions that, when expanded, violate genome boundaries.}
}
\description{
Calculates the number of features within a defined distance of each site.
}
