% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_pois_mix.R
\name{fit_pois_mix}
\alias{fit_pois_mix}
\title{Poisson Mixture Model}
\usage{
fit_pois_mix(x, size = 1e+05)
}
\arguments{
\item{x}{The observed counts}

\item{size}{The number of counts to sample from x.}
}
\description{
Fits a Poisson mixture model to a vector of counts. Parameters are estimated by MLE. Generates initial mean estimates and mixing probabilities by k-means clustering. Returns an n x 2 matrix, where n is the number of distributions, column 1 holds the mixing proportions and column 2 holds the distribution means. The distribution with the highest mean value is returned in returned in row 1. The distribution with the smallest mean value is returned in row n.
}
