% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_width.R
\name{fix_width}
\alias{fix_width}
\title{Fix Coordinate Widths}
\usage{
fix_width(site.list, current.pos = 1, target.pos = 1, genome.obj)
}
\arguments{
\item{site.list}{The list of GRanges objects or GRangesList containing the mapped integration site coordinates.}

\item{current.pos}{The position in the target site duplication described by the current start position. Defaults to 1.}

\item{target.pos}{The desired position of the shifted start position in the target site duplication. Defaults to 1.}

\item{genome.obj}{The BSgenome object of interest.}
}
\value{
A list of GRanges objects or a GRangesList containing coordinates with width = 1.
}
\description{
Shift integration site positions to the desired position in the target site duplication and enforce that their width is 1.
}
