% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_chromosome_seqs.R
\name{get_chromosome_seqs}
\alias{get_chromosome_seqs}
\title{Get Chromosome Sequences}
\usage{
get_chromosome_seqs(genome.obj)
}
\arguments{
\item{genome.obj}{The genome object of interest.}
}
\value{
A list of DNAString objects where each element contains a single chromosome from genome.obj
}
\description{
Extracts the nucleotide sequences of each chromosome in a defined genome object
}
