% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_random_coordinates.R
\name{get_random_coordinates}
\alias{get_random_coordinates}
\title{Get Random Coordinates}
\usage{
get_random_coordinates(
  bam.file,
  paired,
  return.reads = FALSE,
  mapq.cutoff = 0,
  simple.cigar = FALSE,
  tag = "NM",
  tagFilter = list(NM = 0),
  yield.size = NA,
  sort = FALSE,
  prefix,
  write = FALSE
)
}
\arguments{
\item{bam.file}{The bam file from the alignment}

\item{paired}{Boolean. Whether or not the aligned data came from single- or paired-end reads. Defaults to TRUE.}

\item{return.reads}{Boolean. Whether or not to return the start/end coordinates for each read or for each mapped fragment.}

\item{mapq.cutoff}{The desired mapq cutoff value.}

\item{simple.cigar}{Boolean. Identical to simpleCigar in Rsamtools::ScanBamParam().}

\item{tag}{A vector of tags on which to filter the input BAM file. To omit, set tag=character(0).}

\item{tagFilter}{List of possible values for the given tags. To omit entirely, set tagFilter=list().}

\item{yield.size}{The number of reads from the bam file to process at a time.}

\item{sort}{Boolean. Whether or not to sort the generated GRanges object. Defaults to FALSE. Ignores strandedness.}

\item{prefix}{The desired output file prefix. Used to name the output RData file.}

\item{write}{Boolean. Whether or not to save the output to disk.}
}
\value{
A GRanges object containing the uniquely mapped random fragments.
}
\description{
Extracts uniquely mapped random fragments from bam files generated using simulated random data.
Iterates through the bam file in yield.size chunks to save memory.
Outputs the random fragment positions in a GRanges object and saves the object to disk as an RData file.
}
