% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_random_fragments.R
\name{get_unique_random_fragments}
\alias{get_unique_random_fragments}
\title{Get Unique Random Fragments}
\usage{
get_unique_random_fragments(
  bam.file,
  mapq.cutoff,
  yield.size = NA,
  sort = FALSE,
  prefix,
  write = FALSE
)
}
\arguments{
\item{bam.file}{The bam file from the alignment}

\item{mapq.cutoff}{The desired mapq cutoff value. If no quality fltering is desired, set to 0.}

\item{yield.size}{The number of reads from the bam file to process at a time.}

\item{sort}{Logical. Whether or not to sort the generated GRanges object. Defaults to FALSE. Ignores strandedness.}

\item{prefix}{The desired output file prefix. Used to name the output RData file.}

\item{write}{Boolean. Whether or not to save the output to disk.}
}
\value{
A GRanges object containing the uniquely mapped random fragments.
}
\description{
Extracts uniquely mapped random fragments from bam files generated using simulated random data. Iterates through the bam file in yield.size chunks to save memory. Outputs the random fragment positions in a GRanges object and saves the object to disk as an RData file.
}
