% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_comparisons.R
\name{global_comparisons}
\alias{global_comparisons}
\title{Wrap Basic Global Comparisons}
\usage{
global_comparisons(
  xint.obj,
  comparison.type = c("glm", "mean", "wilcoxon", "prop"),
  mean.test = c("t.test", "anova"),
  prop.test = c("fisher", "chi", "G"),
  p.adjust.method = "BH"
)
}
\arguments{
\item{xint.obj}{An xIntOverlap object or a list of xIntOverlap objects (output by bulk_comparisons()).}

\item{comparison.type}{One of 'glm', 'mean', 'wilcoxon', or 'prop'.
Determines the use of overlap_comparisons(), overlap_mean_tests(), wilcox_comparisons(), or overlap_prop_tests().}

\item{mean.test}{One of 't.test', or 'anova'. Only used when comparison.type = 'mean'.}

\item{prop.test}{One of 'fisher', 'chi', or 'G'. Only used when comparison.type = 'prop'.}

\item{p.adjust.method}{The method to use for p-value adjustments. Defaults to "BH".}
}
\value{
A data frame containing effect size metrics and p-values for each comparison.
}
\description{
A basic wrapper for global comparison functions. Outputs a data frame containing effect size metrics and adjusted
p-values. Comparisons made via GLM utilize the quasi-Poisson model family and other default parameters in overlap_comparisons().
This function does not output all of the information returned in the comparison functions themselves.
For more detailed output, run the functions individually.
}
\examples{
data(xobj)
# add examples

}
