% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sites.R
\name{import_sites}
\alias{import_sites}
\title{Import Integration Site Coordinates}
\usage{
import_sites(
  path,
  pattern = ".bed",
  genome.obj = NULL,
  levels.style = NULL,
  keep.metadata = FALSE,
  return.GRangesList = FALSE
)
}
\arguments{
\item{path}{The path of the mapped integration site datasets.}

\item{pattern}{The extension of the integration site dataset files. Defaults to ".bed".}

\item{genome.obj}{The BSgenome object corresponding to genome of interest.
When not NULL, the seqlevels in each IS dataset are compared to the seqlevels of the genome.obj.
Seqlevels present in the IS datasets but not present in the genome.obj are removed.}

\item{levels.style}{The preferred style of seqlevels.}

\item{keep.metadata}{Boolean. Whether or not the keep metadata contained in the imported datasets. Defaults to FALSE.}

\item{return.GRangesList}{Boolean. Whether or not the return the list of imported integration site datasets as a GRangesList. Defaults to FALSE.}
}
\value{
A list of GRanges objects or a GRangesList containing mapped integration sites.
}
\description{
Import a set of integration site datasets from a defined path.
}
