% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_colData.R
\name{join_colData}
\alias{join_colData}
\title{Combine Column Data}
\usage{
join_colData(..., annotations)
}
\arguments{
\item{...}{The xIntObjects of interest.}

\item{annotations}{A character vector containing the names of the feature-sets used to create each xIntObject.}
}
\value{
A data frame containing the colData from each xIntObject.
}
\description{
Combines colData from multiple xIntObjects into a single data frame.
A new 'annotation' column is added to the resulting data frame,
holding the name of the annotations used to create each xIntObject.
}
\examples{
data(xobj)
yobj <- xobj
join_colData(xobj, yobj, annotations = c("Anno1", "Anno2"))

}
