% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo_ci.R
\name{logo_ci}
\alias{logo_ci}
\title{Estimate Logo Credible Intervals}
\usage{
logo_ci(
  sites,
  genome.obj,
  seq.len = 12,
  ignore.strand = FALSE,
  current.start = 1,
  tsd = 5,
  alpha = c(A = 0.3, C = 0.2, G = 0.2, T = 0.3),
  ci = c(0.025, 0.975),
  seq.samp = NULL,
  dir.samp = 5000,
  seed = NULL,
  return.plot = TRUE,
  add.title = FALSE,
  wrap = FALSE,
  ...
)
}
\arguments{
\item{sites}{A SiteList object.}

\item{genome.obj}{The BSgenome object for the genome of interest.}

\item{seq.len}{The desired length of the expanded sequences.}

\item{ignore.strand}{Boolean. Whether or not to ignore strandedness when extracting sequences.
Defaults to FALSE. If TRUE, the sequence from the forward strand is returned.}

\item{current.start}{The position in the target site duplication currently described by the start coordinates in sites.
This is used internally for centering the integration site coordinates.}

\item{tsd}{The total length of the target site duplication.
This is used along with current.start for centering the integration site coordinates.}

\item{alpha}{A scalar or vector of pseudo-counts to add to the observed counts for each nucleotide at each position.
Defaults to c(A = 0.3, C = 0.2, G = 0.2, T = 0.3). Be mindful of order!}

\item{ci}{A numeric vector denoting the upper- and lower-boundaries of the credibile interval. Defaults to c(0.025, 0.975).}

\item{seq.samp}{If not NULL, the number of sites to sample from a given dataset. This can significantly speed up processing time,
at the expense of including every site in the sequence logo. Defaults to NULL}

\item{dir.samp}{The number of samples to draw from the Dirichlet posterior distribution. Defaults to 5000.}

\item{seed}{The seed value for random sampling.}

\item{return.plot}{Boolean.
Whether or not to plot the results. Defaults to TRUE.}

\item{add.title}{Boolean. Whether or not to add the sample name to the plot. Defaults to FALSE.}

\item{wrap}{Boolean. Used with return.plot.
Whether or not to wrap the returned plots into a single output.
When TRUE, arguments ncol and nrow can be used to format the output.}

\item{...}{Used to pass additional plotting parameters to patchwork::wrap_plots().
Examples include ncol and nrow.}
}
\value{
If return.plot = TRUE, a list of ggplot2 objects or a single ggplot2 object containing sequence logos
for the mapped integration site datasets in sites.
If return.plot = FALSE, a list of data frames storing mean and median bit values, along with the estimated CIs.
}
\description{
Estimates credible intervals for each base at each position in a sequence logo.
}
\examples{
library(BSgenome.Hsapiens.UCSC.hs1)
data(sites2)
logo_ci(sites = sites2,
        genome.obj = Hsapiens)

}
