% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fragments.R
\name{make_fragments}
\alias{make_fragments}
\title{Make Fragments}
\usage{
make_fragments(
  insert.sites,
  frag.sites = NULL,
  random = TRUE,
  mean = 500,
  sd = 250,
  genome.obj
)
}
\arguments{
\item{insert.sites}{The generated random integration sites}

\item{frag.sites}{The genomic cleavage sites}

\item{random}{Boolean. Whether or not frag.sites were generated randomly (TRUE) or by simulated restriction digestion (FALSE)}

\item{mean}{Only required when random is TRUE. The target mean fragment size. 500 bp default.}

\item{sd}{Only required when random is TRUE. The target fragment size standard deviation. 250 bp default.}

\item{genome.obj}{The name of the genome object of interest.}
}
\description{
Make random fragments from random integration site positions and genomic cleavage sites
}
