% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_logo.R
\name{make_logo}
\alias{make_logo}
\title{Make Sequence Logos}
\usage{
make_logo(
  site.list,
  seq.len = 50,
  genome.obj,
  current.start = 3,
  tsd = 5,
  return.plot = TRUE
)
}
\arguments{
\item{site.list}{The list or GRangesList containing the mapped site coordinates.}

\item{seq.len}{The length of the expanded sequences.}

\item{genome.obj}{The genome object of interest.}

\item{current.start}{The position in the target site duplication currently described by start. This is used for centering the site coordinates.}

\item{tsd}{The total length of the target site duplication. This is used for centering the site coordinates.}

\item{return.plot}{Boolean. Whether or not to return diagnostic plots for each dataset in site.list instead of the consensus matrix. Defaults to TRUE.}
}
\description{
Generates sequence logos for each dataset in site.list.
}
