% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xInt_dataset.R
\name{make_xInt_dataset}
\alias{make_xInt_dataset}
\title{Make xInt Dataset}
\usage{
make_xInt_dataset(
  site.list,
  features,
  sample.names = NULL,
  conditions,
  condition.levels,
  min.overlap = 1,
  id.col = "name"
)
}
\arguments{
\item{site.list}{The list of integration site datasets}

\item{features}{The feature-set of interest}

\item{min.overlap}{The minimum amount of overlap to count two regions as overlapping}

\item{id.column}{The ID column in the features GRanges object. Default "name".}
}
\description{
Creates a ranged summarized experiment object compatible with other xInt functions. Stores summary information for each sample in column data.
}
