% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xIntObject.R
\name{make_xIntObject}
\alias{make_xIntObject}
\title{Make xIntObject}
\usage{
make_xIntObject(
  site.list,
  features,
  conditions,
  condition.levels,
  min.overlap = 1L,
  id.col = "name"
)
}
\arguments{
\item{site.list}{The list of integration site datasets}

\item{features}{The feature-set of interest.}

\item{conditions}{The infection condition for each dataset. Must be same length as site.list.}

\item{condition.levels}{The factor levels of conditions. Length should be the number of unique conditions.}

\item{min.overlap}{The minimum amount of overlap to count two regions as overlapping}

\item{id.col}{The ID column in the features GRanges object. Default "name".}
}
\value{
A SummarizedExperiment object containing local and global feature overlap information for each dataset in site.list.
}
\description{
Creates a ranged summarized experiment object compatible with other xInt functions. Stores summary information for each sample in column data. Range information is stored in row data.
}
