% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_genome.R
\name{mask_genome}
\alias{mask_genome}
\title{Build Genome Mask}
\usage{
mask_genome(
  aligned.fragments,
  generated.fragments = NULL,
  verify.frags = TRUE,
  genome.obj,
  win.size = 100,
  reltol = 1e-12,
  model = c("pois", "nbinom"),
  ignore.chromosomes = NULL,
  n.cores = 1
)
}
\arguments{
\item{aligned.fragments}{A GRanges object containing the coordinates of the genomic fragments extracted from the alignment BAM file.}

\item{generated.fragments}{A GRanges object containing the coordinates of the genomic fragments generated via \if{html}{\out{<code>}}make_fragments()\if{html}{\out{</code>}}. These represent ground-truth fragments.}

\item{genome.obj}{The BSgenome object for the target genome.}

\item{win.size}{The tile size with which to bin the genome. Defaults to 100.}

\item{reltol}{The relative tolerance for convergence. Used in the Baum-Welch algorithm to fit HMM parameters. Corresponds to the absolute relative difference between iterations. Defaults to 1E-10.}

\item{model}{The model to use in the EM algorithm. The options are \if{html}{\out{<code>}}pois\if{html}{\out{</code>}} and \if{html}{\out{<code>}}nbinom\if{html}{\out{</code>}}}

\item{ignore.chromosomes}{Character vector of chromosome names to ignore. These are not included in the HMM, and should therefore be removed from downstream analyses that depend on the HMM.}

\item{n.cores}{The number of cores to use during coarse mask trimming.}
}
\description{
Masks the target genome using a hidden Markov model according to genome fragmentation and sequencing methods. Assumes counts along genomic bins follow either a Poisson or negative binomial mixture model. Initial model parameters are estimated using one of fit_nb_mix() or fit_pois_mix(), assign_states(), and estimate_trans_mat(). Returns a GRanges object containing reduced genomic coordinates and their assigned state.
}
