% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mstep.nbinom.R
\name{Mstep.nbinom}
\alias{Mstep.nbinom}
\title{Negative Binomial Maximization Function}
\usage{
Mstep.nbinom(x, cond, pm, pn)
}
\arguments{
\item{x}{is a vector of length n containing the univariate observed process. Alternatively, x could be specified as NULL, meaning that the data will be added later (e.g. simulated).}

\item{cond}{is an object created by estep.}

\item{pm}{is a list object containing the (Markov dependent) parameter values associated with the distribution of the observed process.}

\item{pn}{is a list object containing the observation dependent parameter values associated with the distribution of the observed process.}
}
\description{
Maximization algorithm for negative binomial distributions. Modified from a function written by Reed A. Cartwright (https://gist.github.com/reedacartwright/7b5339862c9abd388e10b101bfed6f07).
}
