% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_sequences.R
\name{mutate_sequences}
\alias{mutate_sequences}
\title{Mutate Sequences}
\usage{
mutate_sequences(
  seq,
  mismatch.prob = 0.0025,
  indel.prob = 2.5e-05,
  max.bp = 150
)
}
\arguments{
\item{seq}{The sequence to be mutated.}

\item{mismatch.prob}{The mismatch probability. Defaults to 0.0025.}

\item{indel.prob}{The indel probability. Defaults to 2.5E-5.}

\item{max.bp}{The maximum read length.}
}
\value{
A mutated DNAString object.
}
\description{
Mutate NGS reads at a defined error rate.
}
