% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_random_counts.R
\name{nb_random_counts}
\alias{nb_random_counts}
\title{Fit Random Counts to NB Distribution}
\usage{
nb_random_counts(
  sites,
  mask,
  offset = 0,
  genome.obj = NULL,
  features = NULL,
  size = 1e+06,
  win.size = 100
)
}
\arguments{
\item{sites}{A GRanges object containing the coordinates of the random genomic fragments.}

\item{mask}{A GRanges object containing the masked regions of the genome (e.g. the output of mask_genome()).}

\item{offset}{The nucleotide position of interest relative to the first nucleotide of the target site duplication. Defaults to 0, i.e. the first nucleotide of the duplication. Only used when width of given sites is not 1.}

\item{genome.obj}{The BSgenome object for the target genome.}

\item{size}{The number of observations to use in the fit. Defaults to 1E6.}

\item{win.size}{The tile size with which to bin the genome. Defaults to 100.}
}
\description{
Holder text.
}
