% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_overdispersed_count_tests.R
\name{non_overdispersed_count_tests}
\alias{non_overdispersed_count_tests}
\title{Non-overdispered NHSTs for Count Data}
\usage{
non_overdispersed_count_tests(
  xint.obj,
  refcols = NULL,
  p.adj = "BH",
  labels = NULL
)
}
\arguments{
\item{xint.obj}{The xInt object containing the relevant data.}

\item{refcols}{The columns containing data for the reference samples (e.g. wild-type). Can be a character or numeric vector.}

\item{p.adj}{The method by which to correct p-values for multiple comparisons. Can be NULL. Defaults to "BH". See p.adjust() documentation for more details.}

\item{labels}{A character vector defining output labels for the comparisons. Can be NULL.}
}
\description{
Applies standard non-overdispered NHSTs to an xInt object. The reference group (e.g. wild-type) can be defined via the refcols (reference columns) argument. The function pplies all tests at once. Pairwise tests are run for all samples in the xInt object. Average tests are done for each column (data set) relative to the average of the reference columns (refcols). If reference columns are not specified, the average is taken across all columns. The pairwise tests applied are Fisher's exact test, chi-squared test, and G-test. The average tests are binomial and Poisson exact tests.
}
