% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_comparisons.R
\name{overlap_comparisons}
\alias{overlap_comparisons}
\title{GLM Feature Overlap Comparisons}
\usage{
overlap_comparisons(
  xint.obj,
  min.count,
  min.total.count,
  remove.zeros = TRUE,
  plot = TRUE,
  formula = NULL,
  batch = NULL,
  contrasts = NULL,
  type = c("local", "global"),
  intercept = TRUE,
  model.family = NULL
)
}
\arguments{
\item{xint.obj}{The xInt object of interest.}

\item{min.count}{Used to filter out lowly targeted features.
If not provided, defaults to 1.}

\item{min.total.count}{Used alongside min.count to filter lowly targeted features.
If not provided, defaults to 75\% of the number of samples.}

\item{remove.zeros}{Boolean. Whether or not to remove rows with all zeros from the count matrix.
Defaults to TRUE.}

\item{plot}{Boolean. Whether or not to output diagnostic plots.
Defaults to TRUE.}

\item{formula}{A formula object or character string specifying the model. If NULL (default),
a formula will be constructed based on the condition variable and any provided batch variables.}

\item{batch}{A character vector of column names in the colData of the xIntObject containing relevant
batch information. Defaults to NULL.}

\item{contrasts}{Character vector of contrasts to test, in the form "A-B", denoting pairs and directionality.
If NULL (default), all pairwise comparisons between conditions will be created.}

\item{type}{Character string specifying the type of comparisons: per-feature ('local') or feature-set ('global').
Default 'local'.}

\item{intercept}{Boolean. Whether or not to include an intercept term in the model.
Including an intercept means coefficient effects and significance values are defined relative to a reference
(i.e., the first term in conditions).
Default TRUE.}

\item{model.family}{The model family to use in the GLM. Can be NULL (default), 'poisson', or 'quasipoisson'.
When NULL, model.family is set to 'poisson' when type = 'local' and 'quasipoisson' when type = 'global'.}
}
\value{
A list containing the results of GLM fits for each feature, including:
\item{kept.feats}{Vector of features that passed filtering}
\item{perc.change}{Matrix of percentage changes for pairwise comparisons}
\item{log2.fc}{Matrix of log2 fold changes for pairwise comparisons}
\item{coefficients}{Matrix of coefficient estimates}
\item{p.vals}{Matrix of p-values for each coefficient}
\item{contrasts}{Matrix of log fold changes for pairwise comparisons}
\item{contrast.p.vals}{Matrix of p-values for pairwise comparisons}
\item{model.info}{Information about the model used}
\item{contrast.mat}{The contrast matrix used for comparisons}
}
\description{
Perform per-feature or feature-set integration targeting comparisons using (quasi-)Poisson GLMs.
Supports Poisson and Quasi-Poisson models with flexible formula specification.
Library size offsets are used to account for differences in sequencing depth.
}
\examples{
data(xobj)
local_comparisons(xint.obj = xobj, plot = FALSE)

}
