% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_mean_tests.R
\name{overlap_mean_tests}
\alias{overlap_mean_tests}
\title{Global Overlap Comparisons Using Mean Tests}
\usage{
overlap_mean_tests(
  xint.obj,
  p.adjust.method = "BH",
  method = c("t.test", "anova")
)
}
\arguments{
\item{xint.obj}{The xInt object of interest.}

\item{p.adjust.method}{The p-value adjustment method.
Defaults to "BH". See p.adjust() for more information.}

\item{method}{Method for comparison. One of "t.test" (default) for pairwise t-tests or "anova" for ANOVA followed by Tukey HSD.}
}
\value{
A data frame containing the results of each comparison.
}
\description{
Perform global pairwise comparisons on an xIntOverlap object using tests for equal means.
Comparisons are made using either t-tests or ANOVA followed by Tukey HSD.
}
\examples{
data(xobj)
mean_tests(xint.obj = xobj)

}
