% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_prop_tests.R
\name{overlap_prop_tests}
\alias{overlap_prop_tests}
\title{Global Overlap Comparisons Using Proportion Tests}
\usage{
overlap_prop_tests(
  xint.obj,
  comparison = c("pooled", "all"),
  test = c("fisher", "chi", "G"),
  p.adjust.method = "BH"
)
}
\arguments{
\item{xint.obj}{The xIntObject containing the relevant data.}

\item{comparison}{The extent of the pairwise comparisons.
'all' will compare every dataset to all other datasets.
This can take a long time.
'pooled' (default) pools the information by condition and performs pairwise comparisons across conditions.}

\item{test}{One of 'fisher', 'chi', or 'G'. Defines the statistical test to be performed.}

\item{p.adjust.method}{The method by which to correct p-values for multiple comparisons.
Defaults to "BH". See p.adjust() documentation for more details.}
}
\value{
A matrix of p-values and effect sizes for each pairwise comparison performed.
}
\description{
Perform global pairwise comparisons on an xIntOverlap object using tests for equal proportions.
}
\examples{
data(xobj)
overlap_prop_tests(xint.obj = xobj)

}
