% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_chrom.R
\name{plot_by_chrom}
\alias{plot_by_chrom}
\title{Plot by Chromosome}
\usage{
plot_by_chrom(
  sites,
  feature.list,
  genome.obj,
  conditions = NULL,
  win.size = 1e+06,
  stack.cutoff = 4,
  chr.omit = NULL,
  chr.include = NULL,
  cond.omit = NULL,
  label.size = 0.8
)
}
\arguments{
\item{sites}{A SiteList object.}

\item{feature.list}{A named list of genomic features to plotted.}

\item{genome.obj}{A BSgenome object.}

\item{conditions}{Optional. A vector of conditions.
Must be the same length as sites.
When given, the datasets belonging to each condition are collapsed into a pooled dataset.}

\item{win.size}{The window size for feature density calculations. Defaults to 1E6.}

\item{stack.cutoff}{Defines the maximum number of layers each plotting level can contain.
Defaults to 4.}

\item{chr.omit}{A character vector of chromosomes to omit from the plot. Defaults to c("chrM", "MT").
Must be NULL if chr.include is given.}

\item{chr.include}{A character vector of chromosomes to include in the plot. Defaults to NULL.}

\item{cond.omit}{A vector of conditions to omit from the plot.}

\item{label.size}{The size of the feature labels. Defaults to 0.8.}
}
\value{
A list of plots, invisibly returned.
}
\description{
Generates karyotype plots for genomic regions and features across chromosomes.
}
\examples{
plot_by_chrom(sites, feature.list = list(genes = genes), genome.obj = Hsapiens)

}
