% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot Density}
\usage{
plot_density(
  x,
  stats.df = NULL,
  condition.levels = NULL,
  print.plot = FALSE,
  plot.title = NULL
)
}
\arguments{
\item{x}{The output of feature_density() with options average = TRUE and conditions are given.}

\item{print.plot}{Boolean. Whether or not to print the plot(s). Defaults to FALSE.}

\item{plot.title}{The title of the plot. When NULL, no title is added to the plot.}
}
\value{
A ggplot2 object.
}
\description{
Plot per-condition average feature density.
}
