% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilcox_comparisons.R
\name{plot_lcpm_distributions}
\alias{plot_lcpm_distributions}
\title{Plot log-CPM distributions by condition}
\usage{
plot_lcpm_distributions(lcpm, dat, prior.count)
}
\arguments{
\item{lcpm}{Log-CPM matrix}

\item{dat}{Sample metadata}

\item{prior.count}{The prior.count added to values before log-transformation}
}
\description{
Plot log-CPM distributions by condition
}
