% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metafeature.R
\name{plot_metafeature}
\alias{plot_metafeature}
\title{Plot Meta-Feature Distributions}
\usage{
plot_metafeature(
  x,
  bins = 100,
  order = NULL,
  average.samples = TRUE,
  shade.under = NULL,
  bp.range = NULL,
  y.scale = c("percent", "fraction")
)
}
\arguments{
\item{x}{The output of metafeature().}

\item{bins}{The number of bins to use for binning the data. Defaults to 100.}

\item{order}{The desired order of the data in the figure legend. All items must be present in x$conditions.}

\item{average.samples}{Boolean. Whether or not to plot the mean and standard deviation across replicates.
Defaults to TRUE. When FALSE, replicate data is pooled.}

\item{shade.under}{A character vector of length 1 defining a reference condition to shade under.
Usually used with a random integration control.
Helps highlight deviations from the reference.}

\item{bp.range}{A numeric vector of length 2 defining the basepair range to display (e.g., c(-1000, 1000)).
If NULL, uses relative position (0 to 1).
By setting this, you are implying that all features are the same length.}

\item{y.scale}{One of "percent" (default) or "fraction". Determines the scale of the y-axis.}
}
\value{
A ggplot2 object for plotting meta-feature distributions.
}
\description{
Plots meta-feature distributions using the output of metafeature().
}
