% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_matrix.R
\name{plot_sample_matrix}
\alias{plot_sample_matrix}
\title{Pairwise count matrix plot}
\usage{
plot_sample_matrix(xint.obj, labels = NULL)
}
\arguments{
\item{xint.obj}{The output of make_xInt_dataset().}

\item{labels}{The desired sample labels. Length should equal the number of columns in the count matrix.}
}
\description{
Plots summary information about each of the datasets used to build the xInt model.
}
