% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_tests.R
\name{prop_tests}
\alias{prop_tests}
\title{Proportion Comparisons}
\usage{
prop_tests(xint.obj, comparison = c("all", "pooled"), p.adj = "BH")
}
\arguments{
\item{xint.obj}{The xIntObject containing the relevant data.}

\item{comparison}{The extent of the pairwise comparisons.
\if{html}{\out{<code>}}all\if{html}{\out{</code>}} will compare every dataset to all other datasets.
This can take a long time.
\if{html}{\out{<code>}}pooled\if{html}{\out{</code>}} pools the information by condition and performs pairwise comparisons across conditions.}

\item{p.adj}{The method by which to correct p-values for multiple comparisons.
Defaults to "BH". See p.adjust() documentation for more details.}
}
\value{
A matrix of p-values and effect sizes for each pairwise comparison performed.
}
\description{
Performs global pairwise comparisons on an xInt object using classic tests for equal proportions.
}
\examples{
data(xobj)
prop_tests(xint.obj = xobj, comparison="pooled")

}
