% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_coarse_mask.R
\name{refine_coarse_mask}
\alias{refine_coarse_mask}
\title{Refine Coarse Mask}
\usage{
refine_coarse_mask(coarse.mask, fragments, win.size = 5, n.cores = 1)
}
\arguments{
\item{coarse.mask}{The GRanges object containing the coarsely masked genomic regions.}

\item{fragments}{The GRanges object containing the simulated random fragments/sites.}

\item{win.size}{The window size used to bin the coarse masked regions. Defaults to 10.}

\item{n.cores}{The number of cores to use in parallelization.}
}
\description{
Refines a coarse genomic mask by identifying non-masked sub-regions within the coarsely masked regions via change point estimation. This is generally not intended to be run alone, but is called within mask_genome().
}
