% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_positions.R
\name{relative_positions}
\alias{relative_positions}
\title{Meta-feature Analysis}
\usage{
relative_positions(site.list, features, metadata = NULL, bins = NULL)
}
\arguments{
\item{site.list}{The list or GRangesList containing the mapped site coordinates.}

\item{features}{The features-of-interest. Must be a GRanges object.}

\item{metadata}{The names of metadata columns to keep. Only works when \if{html}{\out{<code>}}bins=NULL\if{html}{\out{</code>}}. That is, when the relative positions of every site is returned.}

\item{bins}{The number of bins to generate across the features. Defaults to NULL (i.e., the relative position of every site is returned).}
}
\description{
Calculates the relative positions of each site along the features-of-interest. Can return the relative positions for each site in each dataset, or the number of sites within defined bins along the features.
}
