% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fastq.R
\name{save_fastq}
\alias{save_fastq}
\title{Save FASTQ}
\usage{
save_fastq(reads, directory.path, prefix, paired = TRUE, compress = TRUE)
}
\arguments{
\item{reads}{The output from trim_seqs().}

\item{directory.path}{The parent directory path.}

\item{prefix}{The file prefix.}

\item{paired}{Boolean. Whether or not to save paired-end fasta files.}

\item{compress}{Boolean. Whether or not to compress the output fasta file.}
}
\value{
None
}
\description{
Saves simulated sequencing reads as R1 and R2 FASTQ files.
}
