% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_random_data.R
\name{simulate_random_data}
\alias{simulate_random_data}
\title{Simulate Random Data}
\usage{
simulate_random_data(
  genome.obj,
  re.sites = NULL,
  cut.after = 1,
  n.sites,
  mean,
  sd,
  min.width = 40,
  max.distance = 1000,
  max.bp = 150,
  iterations = 1,
  n.cores = 1,
  write.ranges = FALSE,
  prefix = NULL,
  directory.path = NULL,
  compress = TRUE,
  collapse = TRUE
)
}
\arguments{
\item{genome.obj}{The name of the genome object of interest.}

\item{re.sites}{A character vector of restriction enzyme recognition sequences.}

\item{cut.after}{A single integer that defines the nucleotide after which the DNA is cleaved. Defaults to 1.}

\item{n.sites}{The number of sites to generate.}

\item{mean}{The target mean fragment size. Only used for random fragmentation.}

\item{sd}{The target fragment size standard deviation. Only used for random fragmentation.}

\item{min.width}{The minimum desired fragment/read length. Default 14 bp.}

\item{max.distance}{The maximum desired inner distance. Default 1000 bp.}

\item{max.bp}{The maximum read length. Default 150 bp.}

\item{iterations}{The number of random fragment datasets to generate.}

\item{n.cores}{The number of cores to use. Enables parallelization.}

\item{write.ranges}{Logical. Whether or not to write out the coordinates of the generated random fragments. Defaults to FALSE.}

\item{directory.path}{The parent directory path.}

\item{compress}{Boolean. Whether or not to compress the output FASTQ files. Defaults to TRUE.}

\item{collapse}{Boolean. Whether or not to collapse the generated fragments for each iteration into a single object. Defaults to TRUE.}
}
\description{
Wrapper function that automatically calls the functions used to generate random genomic fragments. If re.sites is NULL, random fragmentation is assumed. When random fragmentation is assumed, mean and sd must be defined in the function call (default values in make_fragments() are not used). When re.sites is not NULL, the function tries to find re.sites in the chromosome sequences and uses these positions to make fragments. See the individual functions for more information. This function does not use any of the parallelized functions.
}
