% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_overlap_comparisons.R
\name{single_overlap_comparisons}
\alias{single_overlap_comparisons}
\title{Feature Overlap Comparisons for Single Samples}
\usage{
single_overlap_comparisons(
  xint.obj,
  min.count,
  min.total.count,
  remove.zeros = TRUE,
  contrasts = NULL,
  type = "local",
  prior.count = 0.5,
  lib.prior = NULL
)
}
\arguments{
\item{xint.obj}{The xInt object of interest.}

\item{min.count}{Used to filter out lowly targeted features.
If not provided, defaults to 1.}

\item{min.total.count}{Used alongside min.count to filter lowly targeted features.
If not provided, defaults to 75\% of the number of samples.}

\item{remove.zeros}{Boolean. Whether or not to remove rows with all zeros from the count matrix.
Defaults to TRUE.}

\item{contrasts}{Character vector of contrasts to test, in the form "A-B". If NULL (default),
all pairwise comparisons between conditions will be created.}

\item{type}{Character string specifying the type of comparisons: per-feature ('local') or feature-set ('global').
Default 'local'.}

\item{prior.count}{The constant added to count values to avoid taking log2(0).}

\item{lib.prior}{The constant added to library size when calculating log-CPM.
When NULL, lib.prior = prior.count.}

\item{plot}{Boolean. Whether or not to output diagnostic plots.
Defaults to TRUE.}
}
\value{
A list containing the results of Poisson exact tests for each feature, including:
\item{kept.feats}{Vector of features that passed filtering}
\item{perc.change}{Matrix of percentage changes for pairwise comparisons}
\item{log2.fc}{Matrix of log2 fold changes for pairwise comparisons}
\item{p.values}{Matrix of p-values for pairwise comparisons}
\item{contrasts}{Matrix of log fold changes for pairwise comparisons}
}
\description{
Perform per-feature or feature-set integration targeting comparisons using Poisson exact tests.
This function is designed for cases without replicates per condition.
Library size offsets are used to account for differences in sequencing depth.
}
\examples{
data(xobj)
single_comp <- overlap_comparisons_single(xint.obj = xobj, plot = FALSE)

}
