% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SiteList-class}
\alias{SiteList-class}
\title{SiteList-class}
\description{
SiteList-class
}
\examples{
sites <- SiteList(list(GRanges(seqnames="chr1", ranges=IRanges(start=45, end=45))))
validObject(sites)

}
