% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{SiteList}
\alias{SiteList}
\title{Constructor for SiteList}
\usage{
SiteList(sites)
}
\arguments{
\item{sites}{A GRangesList or a list of GRanges objects}
}
\value{
A SiteList object
}
\description{
Constructor for SiteList
}
\examples{
SiteList(list(GRanges(seqnames="chr1", ranges=IRanges(start=45, end=45))))

}
