% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites.R
\docType{data}
\name{sites}
\alias{sites}
\title{Example Integration Site Data}
\format{
An object of class \code{SiteList} of length 10.
}
\source{
Datasets A1-A4 are described in 10.1128/mBio.02254-20.

Datasets B1-B5 were randomly generated using xInt::simulate_random_data().
}
\usage{
sites
}
\value{
A list of GRanges objects.
}
\description{
A simulated experiment mapping integration sites across infection conditions.
}
\keyword{datasets}
