% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sites2.R
\docType{data}
\name{sites2}
\alias{sites2}
\title{Larger Example Data}
\format{
An object of class \code{list} of length 1.
}
\source{
10.1128/mBio.02254-20.
}
\usage{
sites2
}
\value{
A list of GRanges objects.
}
\description{
A larger example of a single IS dataset.
}
\keyword{datasets}
