% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_random_datasets.R
\name{split_random_datasets}
\alias{split_random_datasets}
\title{Split Random Datasets}
\usage{
split_random_datasets(
  ds.list,
  genome.obj,
  n.cores = 1,
  n.split = 5,
  directory.path = "."
)
}
\arguments{
\item{ds.list}{A list of random fragment position datasets.}

\item{genome.obj}{The BSgenome object of interest.}

\item{n.cores}{The number of cores to use.}

\item{n.split}{The number of chromosome subsets to generate.}

\item{directory.path}{The path of the output directory.}
}
\value{
Saves to disk compressed RData objects containing random fragment positions for subsets of chromosomes. Note that all saved objects contain a variable of the same name.
}
\description{
Splits a list a large random fragment position datasets by chromosome subsets and combines matching subsets from all datasets into a single object.
}
