% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_artifacts.R
\name{subtract_artifacts}
\alias{subtract_artifacts}
\title{Subtract Artifact Sites}
\usage{
subtract_artifacts(site.list, artifact.datasets = NULL, artifact.sites = NULL)
}
\arguments{
\item{site.list}{A list of GRanges objects or a GRangesList containing integration site datasets.}

\item{artifact.datasets}{A character vector containing the names of artifactual datasets in site.list
or a numeric vector containing the indices of artifactual datasets in site.list.
These datasets are removed from the output.}

\item{artifact.sites}{A single GRanges object containing the artifactual sites to be removed from
all entries in site.list.}
}
\value{
A list of GRanges objects or a GRangesList containing the subtracted integration site datasets.
}
\description{
Subtract potentially artifactual integration sites from all other datasets.
}
