% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_mask.R
\name{subtract_mask}
\alias{subtract_mask}
\title{Subtract Mask}
\usage{
subtract_mask(reference, mask)
}
\arguments{
\item{reference}{GRanges object corresponding to the regions/coordinates of interest.}

\item{mask}{GRanges object corresponding to the masked regions.}
}
\value{
A GRanges object containing the subtracted coordinates.
}
\description{
Subtract masked regions of the genome from ranges of interest.
Appends all reference metadata columns except the revmap column (if present) to the subtracted ranges.
}
