% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_uninfected_sites.R
\name{subtract_uninfected_sites}
\alias{subtract_uninfected_sites}
\title{Subtract Uninfected Sites}
\usage{
subtract_uninfected_sites(
  site.list,
  uninfected.datasets,
  return.uninfected = FALSE
)
}
\arguments{
\item{site.list}{A list of GRanges objects or a GRangesList containing integration site datasets.}

\item{uninfected.datasets}{A character vector containing the names of uninfected datasets in site.list or a numeric vector containing the indeces of uninfected datasets in site.list. These datasets are removed in the output when return.uninfected = FALSE.}

\item{return.uninfected}{Boolean. Whether or not to return the combined uninfected site coordinates.}
}
\value{
A list of GRanges objects or a GRangesList containing the subtracted integration site datasets.
}
\description{
Subtract integration sites spuriously obtained from uninfected control samples. Uninfected sites are collapsed into a single
}
