% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted_regions.R
\name{targeted_regions}
\alias{targeted_regions}
\title{Unbiased Target Region Identification}
\usage{
targeted_regions(
  sites,
  genome.obj,
  conditions = NULL,
  bin.size = 1000,
  minseglen = 2,
  min.gapwidth = 1,
  p = 2,
  n = 3,
  difference = FALSE,
  BPPARAM = SerialParam(),
  omit = c("chrM", "MT"),
  min.width = 100
)
}
\arguments{
\item{sites}{A SiteList object}

\item{genome.obj}{The genome object being used.}

\item{conditions}{A character vector of conditions corresponding to the individual datasets.
Must be equal to length(sites) and must be in the appropriate order.
When not NULL (default), replicates from the same condition are pooled.}

\item{bin.size}{The bin size resolution. Smaller values offer more granularity but increase noise.
Defaults to 1000.}

\item{minseglen}{The minimum number of consecutive bins required for a region to be called.
Larger sizes can help decrease spurious calls, but at the expense of identifying smaller regions.
Defaults to 10.}

\item{min.gapwidth}{The maximum distance between adjacent regions for them to be merged.
Defaults to 1.}

\item{p}{Savitzky–Golay polynomial order. Must be less-than n. Defaults to 2.}

\item{n}{Savitzky–Golay window size. Must be odd and greater-than p. Defaults to 3.}

\item{difference}{Boolean. Whether or not to return the unique regions for each sample/condition.
Defaults to FALSE.}

\item{omit}{A character vector of chromosomes to omit from the analysis.
Defaults to c("chrM", "MT")}

\item{min.width}{The minimum allowed width of an identified region. Defaults to 100.}

\item{BiocParallel}{parameters. Passed to bplapply().
Defaults to SerialParam() for cross-platform and cluster compatibility.}
}
\value{
A GRanges object containing or GRangesList object containing targeted regions for each sample/condition.
}
\description{
Unbiased Target Region Identification
}
