% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_seqs.R
\name{trim_seqs}
\alias{trim_seqs}
\title{Trim Fragments}
\usage{
trim_seqs(fragments, min.width = 40, max.distance = 1000, max.bp = 150)
}
\arguments{
\item{fragments}{The fragments generated with BSgenome::getSeq().}

\item{min.width}{The minimum desired fragment/read length. Default 14 bp.}

\item{max.distance}{The maximum desired inner distance. Default 1000 bp.}

\item{max.bp}{The maximum read length. Default 150 bp.}
}
\description{
Trims the fragments to simulate the sequencing run (e.g. 150 bp reads). Creates paired end fragments of a defined maximum size and with a defined inner distance.
}
