% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_aligned_fragments.R
\name{verify_aligned_fragments}
\alias{verify_aligned_fragments}
\title{Verify Aligned Fragments}
\usage{
verify_aligned_fragments(generated.fragments, aligned.fragments)
}
\arguments{
\item{generated.fragments}{The fragments generated by \if{html}{\out{<code>}}make_fragments()\if{html}{\out{</code>}}}

\item{aligned.fragments}{The fragments extracted from the alignment BAM file.}
}
\value{
A GRanges object containing the verified aligned fragments.
}
\description{
Verifies that aligned fragments are contained within the ground-truth fragments generated by \if{html}{\out{<code>}}make_fragments()\if{html}{\out{</code>}}. Aligned fragments that are not contained within the generated fragments are discarded, as they are spuriously aligned. Properly aligned fragments are returned.
}
