% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Volcano Plot for Per-feature Comparisons}
\usage{
volcano_plot(
  x,
  contrast,
  p.cutoff = 0.05,
  effect.cutoff = 1,
  p.col = "p.adj",
  effect.col = "log2.fc",
  label.top = 0,
  x.lim = c(-5, 5),
  y.lim = c(0, 6)
)
}
\arguments{
\item{x}{Output of contrast_stats()}

\item{p.cutoff}{p-value cutoff for highlighting significant features}

\item{p.col}{The column corresponding to the effect size value of interest.
One of 'log2.fc' (default) or 'perc.change'}

\item{label.top}{Number of top features to label in the plot.}

\item{x.lim}{x-axis limits}

\item{y.lim}{y-axis limits}

\item{fc.cutoff}{Log fold change cutoff for highlighting features with large effects}

\item{pc.cutoff}{Percent change cutoff for highlight features with large effects}
}
\value{
A ggplot object or a list of ggplot objects (when multiple contrasts are present).
}
\description{
Creates a volcano plot showing -log10(p) vs log2 fold change or percent change versus for a specific contrast.
Takes as input the output of contrast_stats().
}
\examples{
data(xobj)
# add examples

}
