% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{xIntOverlap}
\alias{xIntOverlap}
\title{Constructor for xIntOverlap}
\usage{
xIntOverlap(counts, colData, ...)
}
\arguments{
\item{counts}{A matrix containing integer count data with row names for features and column names for samples.}

\item{colData}{A DataFrame containing the sample metadata with columns: 'sample', 'total.sites',
'overlapping.sites','fraction.overlap', and 'condition'.}

\item{...}{Additional arguments passed to the RangedSummarizedExperiment constructor.}
}
\value{
An xIntOverlap object.
}
\description{
Constructor for xIntOverlap
}
