% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xobj.R
\docType{data}
\name{xobj}
\alias{xobj}
\title{Example xIntOverlap Object}
\format{
An object of class \code{xIntOverlap} with 41488 rows and 9 columns.
}
\source{
Created with make_xIntOverlap().
}
\usage{
xobj
}
\value{
An xIntOverlap object containing feature counts for the ISs used to create 'sites'.
}
\description{
An xIntOverlap object.
}
\keyword{datasets}
