% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_comparison.R
\name{cross_comparison}
\alias{cross_comparison}
\title{Calculate Cross-Sample Overlap}
\usage{
cross_comparison(site.list, unique.sites = TRUE, overlap.coef = TRUE)
}
\arguments{
\item{site.list}{A SiteListObject object.}

\item{unique.sites}{Boolean. Whether or not to enforce unique sites before calculating overlap.
Defaults to TRUE.}

\item{overlap.coef}{Boolean. Whether or not to return the Szymkiewicz–Simpson overlap coefficient.
Defaults to TRUE.}
}
\value{
A matrix of counts denoting the number of overlapping sites between each sample.
}
\description{
Calculates the number of overlapping sites between each sample in a SiteListObject.
}
\examples{
cross_comparison(site.list = sites)

}
