% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_trans_mat.R
\name{estimate_trans_mat}
\alias{estimate_trans_mat}
\title{Estimate Transition Matrix}
\usage{
estimate_trans_mat(x, n.states)
}
\arguments{
\item{x}{The estimated vector of states.}

\item{n.states}{The number of states in the model.}
}
\description{
Generate the initial estimate of the transition matrix for a hidden Markov model. The estimated vector of states can be generated with assign_states().
}
