% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fragments.R
\name{make_fragments}
\alias{make_fragments}
\title{Make Fragments}
\usage{
make_fragments(
  insert.sites,
  frag.sites = NULL,
  random = TRUE,
  mean = 400,
  sd = 100,
  genome.obj,
  to.chr.ends = TRUE,
  U3 = FALSE
)
}
\arguments{
\item{insert.sites}{The generated random integration sites.}

\item{frag.sites}{The positions in the genome where fragmentation occurs.
For random fragmentation, this should be left NULL.
If using restriction enzymes with pattern specificity, this should be the output of digest().}

\item{random}{Boolean. Whether or not frag.sites are to be generated
randomly (TRUE) or by simulated restriction digestion (FALSE)}

\item{mean}{Only required when random is TRUE.
The target mean fragment size. 400 bp default.}

\item{sd}{Only required when random is TRUE.
The target fragment size standard deviation. 100 bp default.}

\item{genome.obj}{The BSgenome object of interest.}

\item{to.chr.ends}{Boolean. Whether or not to treat chromosome ends as capable of generating
potentially mappable fragments. Defaults to TRUE.}

\item{U3}{Boolean. Whether or not mapping is done from the U3 end (5' LTR relative to top-strand).
Defaults to FALSE.}
}
\value{
A GRanges object of fragment ranges.
}
\description{
Make random fragments from random integration site positions and genomic cleavage sites
}
