% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_logo.R
\name{make_logo}
\alias{make_logo}
\title{Make Sequence Logos}
\usage{
make_logo(
  site.list,
  seq.len = 30,
  genome.obj,
  current.start = 3,
  tsd = 5,
  return.plot = TRUE
)
}
\arguments{
\item{site.list}{The list or GRangesList containing the mapped integration site coordinates.}

\item{seq.len}{The desired length of the expanded sequences.}

\item{genome.obj}{The BSgenome object for the genome of interest.}

\item{current.start}{The position in the target site duplication currently described by the start coordinates in site.list. This is used internally for centering the integration site coordinates.}

\item{tsd}{The total length of the target site duplication. This is used along with current.start for centering the integration site coordinates.}

\item{return.plot}{Boolean. Whether or not to return logos for each dataset in site.list instead of the consensus matrix. Defaults to TRUE.}
}
\value{
If return.plot = TRUE, a list of ggplot2 objects containing sequence logos for the mapped integration site datasets in site.list. If return.plot = FALSE, a list of consensus matrices for each integraiton site dataset.
}
\description{
Generates sequence logos or consensus matrices for integration site datasets.
}
