% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_genome.R
\name{mask_genome}
\alias{mask_genome}
\title{Create Genome Mask}
\usage{
mask_genome(
  mapped.positions,
  genome.obj,
  ignore.chromosomes = NULL,
  n.cores = 1
)
}
\arguments{
\item{mapped.positions}{A GRanges object containing the positions of mapped fragments.}

\item{genome.obj}{The BSgenome object of interest.}

\item{ignore.chromosomes}{Chromosomes to be ignored in masking.
These chromosomes are not included in the final output.}

\item{n.cores}{The number of cores to use for parallel computing.}
}
\value{
A GRanges object containing the masked regions of the genome.
}
\description{
Enumerates regions of the genome that cannot be uniquely mapped
given a particular experimental approach (genome fragmentation method, read length, etc.).
}
