% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeature.R
\name{metafeature}
\alias{metafeature}
\title{Meta-Feature Analysis}
\usage{
metafeature(sites, features, collapse = TRUE, conditions = NULL)
}
\arguments{
\item{sites}{The list or GRangesList containing IS coordinates.}

\item{features}{The features-of-interest. Must be a GRanges object.}

\item{collapse}{Boolean. Whether or not to collapse the output into a single data frame.
Defaults to TRUE. For compatibility with plot_metafeature(), this must be TRUE.}

\item{conditions}{The condition for each dataset. Must be same length as sites.}
}
\value{
A data frame or a list of data frames.
}
\description{
Calculates the relative position of sites along defined features-of-interest.
Returns the relative positions for each site in each dataset.
}
\examples{
data(sites2)
data(xobj)
feats <- rowRanges(xobj)

metafeature(sites = sites2,
            features = feats)

}
