% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_prop_tests.R
\name{overlap_prop_tests}
\alias{overlap_prop_tests}
\title{Global Overlap Comparisons Using Proportion Tests}
\usage{
overlap_prop_tests(xint.obj, comparison = c("pooled", "all"), p.adj = "BH")
}
\arguments{
\item{xint.obj}{The xIntObject containing the relevant data.}

\item{comparison}{The extent of the pairwise comparisons.
'all' will compare every dataset to all other datasets.
This can take a long time.
'pooled' (default) pools the information by condition and performs pairwise comparisons across conditions.}

\item{p.adj}{The method by which to correct p-values for multiple comparisons.
Defaults to "BH". See p.adjust() documentation for more details.}
}
\value{
A matrix of p-values and effect sizes for each pairwise comparison performed.
}
\description{
Perform global pairwise comparisons on an xIntOverlap object using tests for equal proportions.
}
\examples{
data(xobj)
prop_tests(xint.obj = xobj, comparison="pooled")

}
