% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot Density}
\usage{
plot_density(
  x,
  stats.df = NULL,
  condition.levels = NULL,
  print.plot = FALSE,
  y.label = "Genes / Mb",
  numeric.p = FALSE,
  seed = 1,
  add.legend = TRUE
)
}
\arguments{
\item{x}{The output of feature_density() with options average = TRUE and conditions are given.}

\item{stats.df}{The output of compare_density().}

\item{condition.levels}{The condition levels for the samples. If NULL (default), the levels are created in the order they appear.}

\item{print.plot}{Boolean. Whether or not to print the plot(s). Defaults to FALSE.}

\item{y.label}{The label of the y-axis. Defaults to "Genes / Mb".}

\item{numeric.p}{Boolean. When FALSE (default), significance is denoted symbolically.
p > 0.05 is 'ns', 0.01 >= p < 0.05 is '\emph{', 0.001 >= p < 0.01 is '\strong{', and p < 0.001 is '}}'.
When TRUE, the actual p-values are given.}

\item{seed}{Sets the seed in position_jitter(). Defaults to 1.}

\item{add.legend}{Boolean. Whether or not to include a legend on the plots. Defaults to TRUE.}
}
\value{
A ggplot2 object.
}
\description{
Plot per-condition average feature density.
}
