% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_global_overlap.R
\name{plot_global_overlap}
\alias{plot_global_overlap}
\title{Plot Feature-Set Overlap}
\usage{
plot_global_overlap(
  xint.obj,
  stats.df = NULL,
  numeric.p = FALSE,
  seed = 1,
  print.plot = FALSE,
  add.title = FALSE,
  wrap = FALSE,
  add.legend = FALSE,
  plot.titles = NULL,
  y.scale = c("percent", "fraction"),
  ...
)
}
\arguments{
\item{xint.obj}{An xIntOverlap object.}

\item{stats.df}{A data frame containing significance values for each comparison.
Must have the columns 'comparison' and 'p.adj' (even if the p-values are not adjusted).
The outputs of e.g., global_comparisons(), overlap_comparisons() with type = 'global',
and overlap_mean_tests() will all work.}

\item{numeric.p}{Boolean. When FALSE (default), significance is denoted symbolically.
p > 0.05 is 'ns', 0.01 >= p < 0.05 is '\emph{', 0.001 >= p < 0.01 is '\strong{', and p < 0.001 is '}}'.
When TRUE, the actual p-values are given.}

\item{seed}{Sets the seed in position_jitter(). Defaults to 1.}

\item{print.plot}{Boolean. Whether or not to print the plot(s). Defaults to FALSE.}

\item{add.title}{Boolean. Whether or not to add the name of the genomic feature to the plot. Defaults to FALSE.}

\item{wrap}{Boolean. Whether or not to wrap a list of plots into a single output. Defaults to FALSE.}

\item{add.legend}{Boolean. Whether or not to include a legend on the plots. Defaults to FALSE.}

\item{plot.titles}{A character vector containing the names of plot titles. Only used with stats.df is NULL and add.titles is TRUE.}
}
\value{
A ggplot2 object.
}
\description{
Plots global feature-set overlap for each condition in an xIntOverlap object.
}
