% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sites.R
\name{random_sites}
\alias{random_sites}
\title{Generate Random Sites}
\usage{
random_sites(n.sites, genome.obj)
}
\arguments{
\item{n.sites}{The number of sites to generate.}

\item{genome.obj}{The BSgenome object of interest.}
}
\value{
A GRanges object containing the integration site positions.
}
\description{
Generates randomly positioned integration sites within the genome of interest.
}
