% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_positions.R
\name{relative_positions}
\alias{relative_positions}
\title{Relative Positions Along Features}
\usage{
relative_positions(site.list, features, bins = NULL, metadata = NULL)
}
\arguments{
\item{site.list}{The list or GRangesList containing the integration site coordinates.}

\item{features}{The features-of-interest. Must be a GRanges object.}

\item{bins}{The number of bins to generate across the features. Defaults to NULL, which returns the relative position of every site..}

\item{metadata}{The names of metadata columns to keep. Only used when the data are not binned.}

\item{ignore.strand}{Boolean. Whether or not to return relative positions relative to feature directionality. Defaults to FALSE (directionality is considered). This has no influence on overlap quantification.}
}
\value{
A list of data frames.
}
\description{
Calculates the relative position of each integration site along the features-of-interest. Returns either the relative positions for each site in each dataset or the number of sites within defined bins along the features.
}
