% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fasta.R
\name{save_fasta}
\alias{save_fasta}
\title{Save FASTA}
\usage{
save_fasta(reads, directory.path, prefix, compress = TRUE)
}
\arguments{
\item{reads}{The output from trim_seqs().}

\item{directory.path}{The parent directory path.}

\item{prefix}{The file prefix.}

\item{compress}{Logical. Whether or not to compress the output fasta file.}
}
\value{
None
}
\description{
Saves simulated sequencing reads as R1 and R2 fasta files.
}
