% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_artifacts.R
\name{subtract_artifacts}
\alias{subtract_artifacts}
\title{Subtract Artifact Sites}
\usage{
subtract_artifacts(sites, artifact.datasets = NULL, artifact.sites = NULL)
}
\arguments{
\item{sites}{A SiteList object.}

\item{artifact.datasets}{A character vector containing the names of artifactual datasets in sites
or a numeric vector containing the indices of artifactual datasets in sites.
These datasets are removed from the output.}

\item{artifact.sites}{A single GRanges object containing the artifactual sites to be removed from
all entries in sites.}
}
\value{
A list of GRanges objects or a GRangesList containing the subtracted integration site datasets.
}
\description{
Subtract potentially artifactual integration sites from all other datasets.
}
