% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xobj.R
\docType{data}
\name{xobj}
\alias{xobj}
\title{Example xIntObject}
\format{
An object of class \code{xIntObject} with 0 rows and 9 columns.
}
\source{
Created with make_xIntObject().
}
\usage{
xobj
}
\value{
An xIntObject containing feature counts for the ISs used to create 'sites'.
}
\description{
An xIntObject.
}
\keyword{datasets}
